[ADF-3336] fixed selectors and check for document list test
